ajout constructeur dans player DAO Correction DAOMatch find Correction DAOcourt Correctin GamesDAO Correction setDAO CREATE Correction dbo Game au lieu de dbo games

+++ DAO/CourtDAO.cs
@@ -68,7 +68,7 @@ namespace Tennis_Tournament_Console.DAO
                     {
                         court.setId(reader.GetInt32(0));
                         court.setNbSpectators(reader.GetInt32(1));
-                        court.setCovered(reader.GetByte(2) != 0);
+                        court.setCovered(reader.GetInt32(2) != 0);
                     }
                 }
             }
